static-cptr-cache: Adds lock to cptr cache, fixes race conditions in tests.
All looks OK now.

Few other misc things:

   - config.h header goes with install
   - Some debug code in cap.c.
   - Wasn't handling return value of make cnode table in cap.c.

For now, until we (possibly) move stuff around, I'm including the
internal header in the public header so that I have access to
the cap mutex type. I'm considering re-working things so that this
won't be necessary in the future. Temporary hack for now.