Move to using a prefixed config.h .
This library has consumers that build against its installed
version.  For that, and for the other general problems that
including config.h in public headers causes, just make config.h be
prefixed so it is "unique" and install that one.

It was either this, or hack the headers once again with some
foo like that).  But let's not do that now... the headers need a reorg
before something like that happens.