private-types: Adds struct cap_type_system. Old code builds.
The code builds right now by default (--enable-global-cap-types is
on by default). Next step is to adjust the internals.

Fixed an autoconf script bug: if you AC_SUBST a variable in one
case and AC_SUBST_FILE in another (perhaps in an if statement), autoconf
gets confused. So, I just always AC_SUBST now and cat the file into
the variable when necessary. (I'm referring to the
CAP_INCLUDE_GLOBAL_TYPES variable in configure.ac.)