Fix kernel external build.
The kernel build is a bit fragile right now (at least the kernel
module build is). In order for all kernel build targets to
work properly (e.g. make clean), the Kbuild file needs to be
at the "root" of the build tree. For example, the kernel's make
clean uses a simple "find" piped into xargs rm -f, so it only trickles
down child directories. (If you build files that are not a child
of the root Kbuild directory, make clean / find won't touch
them.) This means that a "make clean" won't actually do a full
clean up.