Add padding to cptr_cache.
I want non-isolated code to be able to initialize a cptr_cache
for isolated code. The problem is, the size of a mutex for
isolated code is only 8 bytes, versus 16 bytes for non-isolated.
So, the isolated libcap is compiled so that there is 8 bytes of
padding after the cptr cache lock. I know, this is really awesome ;)