Bug in trylock definition for kernel platform.
Unlike other kernel functions, mutex_trylock returns 1 on sucess
and 0 on failure. We were negating that, and then negating it
again in our check in cap.c. We got away with this in the microkernel
because in the host kernel, the first call returned 1 (for success)
and then 0 (for error); the double negations mixed things up and
it still worked. But inside LCDs, since I hack trylock to always
return 1, it caused the code to go into an infinite loop.

Gonna make this same change and put it in the mainline.