lcd: Add support for clang builds.
This is kind of hack-ish for now. clang doesn't support some of the
compiler flags that are used in a kernel build. I'm providing a
general "hook" that a libcap builder can use to filter out bad
kbuild cflags. (A perhaps better long term solution would check that
the CC supports certain flags, and then do the filtering.)