Fix cap_revoke to work on depth > 2 CDTs
The old algorithm would ignore 2nd-level children of the revoked node
because they were inserted "before" their parent that was getting
removed from the list.

The new algorithm re-starts the revoke from the last "checkpoint" (node
that is guaranteed to be in the output tree) so that we can detect the
children of deleted nodes.